Rename libs to keep with standard
