+++ flexget/ui/plugins/history/static/history.html
