+++ flexget/ui/plugins/schedule/static/schedule.html
