+++ flexget/ui/static/libs/bootstrap/fonts/glyphicons-halflings-regular.eot
