+++ flexget/ui/static/libs/bootstrap/fonts/glyphicons-halflings-regular.woff2
