+++ flexget/ui/static/libs/font-awesome/fonts/fontawesome-webfont.woff2
