+++ flexget/ui/static/libs/schema-form/js/ObjectPath.js
