+++ flexget/ui/static/libs/schema-form/js/bootstrap-decorator.js
