+++ flexget/ui/static/libs/schema-form/js/schema-form.js
